created vote_controller and moved cast_vote, cast_votes, show_votes, refresh_vote_cache out of observer_controller.
