+++ app/controllers/observer_controller.rb
@@ -29,10 +29,6 @@
 # Is this just a private helper?
 #  show_selected_observations:: (helper)
 
-# TODO: Create VoteController with:
-#  cast_vote::
-#  show_votes::
-
 # TODO: Create NotificationController with:
 #  show_notifications::
 #  list_notifications::
@@ -163,7 +159,6 @@ class ObserverController < ApplicationController
     :show_rss_log,
     :show_site_stats,
     :show_user,
-    :show_votes,
     :test,
     :textile,
     :textile_sandbox,
@@ -185,7 +180,6 @@ class ObserverController < ApplicationController
     :show_obs,
     :show_observation,
     :show_user,
-    :show_votes
   ]
 
   ##############################################################################
@@ -1312,63 +1306,6 @@ class ObserverController < ApplicationController
     end
     # render(text: "", layout: true)
     redirect_with_query(action: "show_observation", id: id)
-  end
-
-  ##############################################################################
-  #
-  #  :section: Vote Stuff
-  #
-  ##############################################################################
-
-  # Create vote if none exists; change vote if exists; delete vote if setting
-  # value to -1 (owner of naming is not allowed to do this).
-  # Linked from: (nowhere)
-  # Inputs: params[]
-  # Redirects to show_observation.
-  def cast_vote # :norobots:
-    pass_query_params
-    naming = Naming.find(params[:id].to_s)
-    observation = naming.observation
-    value = params[:value].to_i
-    observation.change_vote(naming, value)
-    redirect_with_query(action: "show_observation", id: observation.id)
-  end
-
-  # This is the new POST method for show_observation.
-  def cast_votes # :norobots:
-    pass_query_params
-    observation = find_or_goto_index(Observation, params[:id].to_s)
-    return unless observation
-    if params[:vote]
-      flashed = false
-      observation.namings.each do |naming|
-        value = param_lookup([:vote, naming.id.to_s, :value], &:to_i)
-        next unless value &&
-                    observation.change_vote(naming, value) &&
-                    !flashed
-        flash_notice(:runtime_show_observation_success.t)
-        flashed = true
-      end
-    end
-    redirect_with_query(action: "show_observation", id: observation.id)
-  end
-
-  # Show breakdown of votes for a given naming.
-  # Linked from: show_observation
-  # Inputs: params[:id] (naming)
-  # Outputs: @naming
-  def show_votes # :nologin: :prefetch:
-    pass_query_params
-    @naming = find_or_goto_index(Naming, params[:id].to_s)
-  end
-
-  # Refresh vote cache for all observations in the database.
-  def refresh_vote_cache # :root: :norobots:
-    return unless is_in_admin_mode?
-    # Naming.refresh_vote_cache
-    Observation.refresh_vote_cache
-    flash_notice(:refresh_vote_cache.t)
-    redirect_to(action: "list_rss_logs")
   end
 
   ##############################################################################
