+++ app/views/naming/_show.html.erb
@@ -6,7 +6,7 @@
 
 <!--TABLE OF NAMES FOR WIDE SCREEN-->
 <div class="hidden-xs">
-  <%= form_tag(add_query_param(controller: :observer, action: :cast_votes,
+  <%= form_tag(add_query_param(controller: :vote, action: :cast_votes,
                id: observation.id), id: "cast_votes_1") do %>
 
     <table class="table-namings">
@@ -42,7 +42,7 @@
             </td>
             <td>
               <% txt = naming.vote_percent.round.to_s.html_safe + "%" %>
-              <%= link_with_query(h(txt), { controller: :observer,
+              <%= link_with_query(h(txt), { controller: :vote,
                                             action: :show_votes,
                                             id: naming.id },
                                   { data: { role: "open_popup",
@@ -125,7 +125,7 @@
 
 <!--TABLE OF NAMES FOR MOBILE-->
 <div class="visible-xs">
-  <%= form_tag(add_query_param(controller: :observer, action: :cast_votes,
+  <%= form_tag(add_query_param(controller: :vote, action: :cast_votes,
                id: observation.id), id: "cast_votes_2") do %>
 
     <div class="row visible-xs">
@@ -159,7 +159,7 @@
                   </div>
                   <div class="col-xs-6">
                     <% txt = naming.vote_percent.round.to_s.html_safe + "%" %>
-                    <%= link_with_query(h(txt), { controller: :observer,
+                    <%= link_with_query(h(txt), { controller: :vote,
                                                   action: :show_votes,
                                                   id: naming.id },
                                                 { data: { role: "open_popup",
@@ -247,7 +247,7 @@
     @naming = naming %>
     <div class="popup" id="show_votes_<%= naming.id %>" data-role="popup">
       <div class="popup-frame" id="show_votes_<%= naming.id %>_frame">
-        <%= render(partial: "observer/show_votes",
+        <%= render(partial: "vote/show_votes",
                    locals: { do_cancel: true, naming: @naming }) %>
       </div>
     </div>
