+++ app/views/vote/show_votes.html.erb
